python/samples/requirements.txt
python-dotenv
soundfile
numpy
scipy
azure-ai-translation-text
azure-cognitiveservices-speech
azure-core
langchain-core
langchain-ollama
langchain-openai
langsmith
ollama
PyAudio
pyttsx3
pyusb
wavio
yacman
sounddevice
openai
matplotlib
numpy
transformers
simpleaudio
torch
